Refactor engine logic
